requirements.txt
asgiref==3.2.3
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
Django==2.2.9
django-background-tasks==1.2.5
django-compat==1.0.15
django-js-asset==1.2.2
django-mptt==0.10.0
django-mssql-backend==2.5.0
django-silk==4.0.0
entrypoints==0.3
flake8==3.7.9
gprof2dot==2019.11.30
gunicorn==20.0.4
idna==2.8
Jinja2==2.10.3
lxml==4.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.6
pkg-resources==0.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
pyodbc==4.0.27
python-dateutil==2.8.1
python-dotenv==0.10.5
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==5.0.1
